Port away from KDELibs4Support, use Solid::Power interface
Summary:
The Solid::Power implementation does not track the state itself (to
avoid querying the initial state even when it is not used), so track
the state inside PkUpdates and query the initial state from the interface
asynchronously.

Test Plan:
The initial state is printed correctly in the debug output, same for
change notifications.

Depends on D21975

Reviewers: lukas, jgrulich

Reviewed By: jgrulich

Differential Revision: https://phabricator.kde.org/D21976